+++ src/main.tsx
@@ -14,7 +14,6 @@ createRoot(document.getElementById('root')!).render(
           <App />
         </EquipmentProvider>
       </UserProvider>
-
     </AuthProvider>
   </StrictMode>,
 )
